Move quiz logic down into Quiz from QuizController
