+++ .eslintrc.js
@@ -18,5 +18,6 @@ module.exports = {
         "react/prefer-stateless-function": "off",
         "react/no-access-state-in-setstate": "off",
         "react/jsx-props-no-spreading": "off",
+        "object-shorthand": "off"
     },
 }
